+++ ChemSharp.Molecules/Properties/Distance.cs
@@ -4,7 +4,7 @@ namespace ChemSharp.Molecules.Properties
 {
     public class Distance : KeyValueProperty
     {
-        public override float Value => MathV.Distance(Atom1.Location, Atom2.Location);
+        public override double Value => MathV.Distance(Atom1.Location, Atom2.Location);
         public override string Key => $"{Atom1.Title} - {Atom2.Title}";
         public override string Unit => "Å";
